refactor: Move component files to appropriate folders
Move components from top $lib to $lib/components, and those components related to UI into $lib/components/ui. This should help code organization
