+++ src/lib/components/PermissionsPage.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
     import Button from "$lib/components/ui/button/button.svelte";
     import { type RecordingStates } from "$lib/types";
-    import { NotificationSystem } from "./notificationSystem.svelte";
-    import Status from "./components/Status.svelte";
+    import { NotificationSystem } from "$lib/notificationSystem.svelte";
+    import Status from "$lib/components/ui/Status.svelte";
 
     interface PermissionsPageProps {
         setupRecorder: () => Promise<void>;
