+++ src/lib/components/WhisperOptions.svelte
